+++ src/newgroups.rs
@@ -13,8 +13,8 @@ use tokio::io::{split, AsyncBufReadExt, AsyncWriteExt, BufReader, BufWriter};
 #[derive(Debug)]
 pub struct Newsgroup {
     pub name: String,
-    pub high: u32,
-    pub low: u32,
+    pub high: u64,
+    pub low: u64,
     pub status: String,
 }
